Adding SSL protection
Client Gui send message to the node.js server by using SSL encryption. Then the message is sending from the node.js server to the boot spring server. They are running on the same computer so this sending is secured and this is why we are not using a SSL encryption for sending this messsage.